# main.go
package main

import (
	"bytes"
	"encoding/json"
	"fmt"
	"github.com/go-chi/chi/v5"
	"github.com/go-chi/chi/v5/middleware"
	"github.com/go-chi/jwtauth/v5"
	httpSwagger "github.com/swaggo/http-swagger"
	"golang.org/x/crypto/bcrypt"
	"log"
	"net/http"
	"strings"
	_ "studentgit.kata.academy/alikkamalrus_gmail.com/Geoservice/docs" // docs is generated by Swag CLI, you have to import it.
	"sync"
	"time"
)

// @title GeoService API
// @version 1
// @description This is a simple GeoService API

// @host localhost:8080
// @BasePath /api

// @securityDefinitions.apikey BearerAuth
// @in header
// @name Authorization

type User struct {
	Username string
	Password string
}

var users = make(map[string]User)
var mu sync.Mutex

var tokenAuth *jwtauth.JWTAuth = jwtauth.New("HS256", []byte("secret"), nil)

type RegisterRequest struct {
	Username string `json:"username"`
	Password string `json:"password"`
}

// @Summary Register new user
// @Description Creates a new user account
// @Tags auth
// @Accept application/json
// @Produce application/json
// @Param request body RegisterRequest true "User credentials"
// @Success 201 {string} string "User created successfully"
// @Failure 400 {string} string "Invalid request format or user already exists"
// @Failure 500 {string} string "Failed to hash password"
// @Router /register [post]
func RegisterHandler(w http.ResponseWriter, r *http.Request) {
	var req struct {
		Username string `json:"username"`
		Password string `json:"password"`
	}

	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, "Invalid request format", http.StatusBadRequest)
		return
	}

	mu.Lock()
	defer mu.Unlock()

	if _, exists := users[req.Username]; exists {
		http.Error(w, "User already exists", http.StatusBadRequest)
		return
	}

	hashedPassword, err := bcrypt.GenerateFromPassword([]byte(req.Password), bcrypt.DefaultCost)
	if err != nil {
		http.Error(w, "Failed to hash password", http.StatusInternalServerError)
		return
	}

	users[req.Username] = User{Username: req.Username, Password: string(hashedPassword)}

	w.WriteHeader(http.StatusCreated)
	w.Write([]byte("User created successfully"))
}

type LoginRequest struct {
	Username string `json:"username"`
	Password string `json:"password"`
}

// @Summary User login
// @Description Authenticates user and returns JWT token
// @Tags auth
// @Accept application/json
// @Produce application/json
// @Router /login [post]
// @Param request body LoginRequest true "User credentials"
// @Success 200 {object} map[string]string "token"
// @Failure 400 {string} string "Invalid username or password"
func LoginHandler(w http.ResponseWriter, r *http.Request) {
	var req struct {
		Username string `json:"username"`
		Password string `json:"password"`
	}

	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, "Invalid request format", http.StatusBadRequest)
		return
	}

	mu.Lock()
	user, exists := users[req.Username]
	mu.Unlock()

	if !exists || bcrypt.CompareHashAndPassword([]byte(user.Password), []byte(req.Password)) != nil {
		http.Error(w, "Invalid username or password", http.StatusBadRequest)
		return
	}

	_, tokenString, _ := tokenAuth.Encode(map[string]interface{}{
		"username": req.Username,
		"exp":      time.Now().Add(time.Hour * 48).Unix(),
	})

	resp := map[string]string{"token": tokenString}
	json.NewEncoder(w).Encode(resp)
}

func AuthMiddleware(next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		_, claims, err := jwtauth.FromContext(r.Context())
		if err != nil {
			http.Error(w, "Unauthorized", http.StatusForbidden)
			return
		}

		fmt.Println("Authenticated user:", claims["username"])
		next.ServeHTTP(w, r)
	})
}

func JWTAuthMiddleware(next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		authHeader := r.Header.Get("Authorization")
		if authHeader != "" && !strings.HasPrefix(authHeader, "Bearer ") {
			r.Header.Set("Authorization", "Bearer "+authHeader)
		}
		next.ServeHTTP(w, r)
	})
}

const dadataURL = "https://suggestions.dadata.ru/suggestions/api/4_1/rs/suggest/address"

type RequestAddressSearch struct {
	Query string `json:"query"`
}

type RequestAddressGeocode struct {
	Lat string `json:"lat"`
	Lng string `json:"lng"`
}

type Address struct {
	City   string `json:"city"`
	Street string `json:"street"`
}

type ResponseAddress struct {
	Addresses []*Address `json:"addresses"`
}

func HomeHandler(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusOK)
	w.Write([]byte("Гео сервис запущен!"))
}

// @Summary Search address
// @Description Search address by query
// @Tags address
// @Accept json/apllication
// @Produce json/apllication
// @Param query body RequestAddressGeocode true "Query"
// @Success 200 {object} ResponseAddress
// @Router /address/geocode [post]
// @Failure 400 {string} string "Invalid request format"
// @Failure 500 {string} string "Service unavailable"
// @Security BearerAuth
func AddressGeocodeHandler(w http.ResponseWriter, r *http.Request) {
	if r.Method != http.MethodPost {
		http.Error(w, "Метод запроса не POST", http.StatusMethodNotAllowed)
		return
	}

	var req RequestAddressGeocode
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, "Invalid request format", http.StatusBadRequest)
		return
	}

	result, err := GeocodeAddress(req.Lat, req.Lng)
	if err != nil {
		http.Error(w, "Service unavailable", http.StatusInternalServerError)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(result)
}

func GeocodeAddress(lat, lng string) (*ResponseAddress, error) {
	dadataAPIKey := "d0e0b799107963c33b1f6e8cd96547b73b46bb16"

	reqBody, _ := json.Marshal(map[string]string{
		"lat": lat,
		"lon": lng,
	})

	req, err := http.NewRequest("POST", "https://suggestions.dadata.ru/suggestions/api/4_1/rs/geolocate/address", bytes.NewBuffer(reqBody))
	if err != nil {
		return nil, err
	}

	req.Header.Set("Content-Type", "application/json")
	req.Header.Set("Accept", "application/json")
	req.Header.Set("Authorization", "Token "+dadataAPIKey)

	client := &http.Client{}
	resp, err := client.Do(req)
	if err != nil {
		return nil, err
	}
	defer resp.Body.Close()

	if resp.StatusCode != http.StatusOK {
		return nil, fmt.Errorf("failed to fetch data from DaData, status code: %d", resp.StatusCode)
	}

	var rawResponse struct {
		Suggestions []struct {
			Data Address `json:"data"`
		} `json:"suggestions"`
	}

	if err := json.NewDecoder(resp.Body).Decode(&rawResponse); err != nil {
		return nil, err
	}
	var response ResponseAddress
	for _, suggestion := range rawResponse.Suggestions {
		response.Addresses = append(response.Addresses, &Address{
			City:   suggestion.Data.City,
			Street: suggestion.Data.Street,
		})
	}

	return &response, nil
}

// @Summary Search address
// @Description Search address by query
// @Tags address
// @Accept application/json
// @Produce application/json
// @Param query body RequestAddressSearch true "Query"
// @Success 200 {object} ResponseAddress
// @Failure 400 {string} string "Invalid request format"
// @Failure 500 {string} string "Service unavailable"
// @Router /address/search [post]
// @Security BearerAuth
func AddressSearchHandler(w http.ResponseWriter, r *http.Request) {
	if r.Method != http.MethodPost {
		http.Error(w, "Метод не POST", http.StatusMethodNotAllowed)
		return
	}

	var req RequestAddressSearch
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, "Invalid request format", http.StatusBadRequest)
	}
	fmt.Println("Request body:", req)

	result, err := SearchAddress(req.Query)
	if err != nil {
		http.Error(w, "Service unavailable", http.StatusInternalServerError)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(result)
}

func SearchAddress(query string) (*ResponseAddress, error) {
	dadataToken := "d0e0b799107963c33b1f6e8cd96547b73b46bb16"

	reqBody, _ := json.Marshal(map[string]string{"query": query})

	req, err := http.NewRequest("POST", dadataURL, bytes.NewBuffer(reqBody))
	if err != nil {
		return nil, err
	}

	req.Header.Set("Authorization", "Token "+dadataToken)
	req.Header.Set("Content-Type", "application/json")

	client := &http.Client{}
	resp, err := client.Do(req)
	if err != nil {
		return nil, err
	}
	defer resp.Body.Close()

	if resp.StatusCode != http.StatusOK {
		return nil, fmt.Errorf("failed to fetch data from DaData, status code: %d", resp.StatusCode)
	}

	var rawResponse struct {
		Suggestions []struct {
			Data Address `json:"data"`
		} `json:"suggestions"`
	}

	if err := json.NewDecoder(resp.Body).Decode(&rawResponse); err != nil {
		return nil, err
	}

	var response ResponseAddress
	for _, suggestion := range rawResponse.Suggestions {
		response.Addresses = append(response.Addresses, &Address{
			City:   suggestion.Data.City,
			Street: suggestion.Data.Street,
		})
	}

	fmt.Println("Response body:", response)
	return &response, nil
}

func main() {
	r := chi.NewRouter()

	r.Use(middleware.Logger)
	r.Use(middleware.Recoverer)

	r.Post("/api/register", RegisterHandler)
	r.Post("/api/login", LoginHandler)

	r.Group(func(protected chi.Router) {
		protected.Use(JWTAuthMiddleware)
		protected.Use(jwtauth.Verifier(tokenAuth))
		protected.Use(AuthMiddleware)

		protected.Post("/api/address/search", AddressSearchHandler)
		protected.Post("/api/address/geocode", AddressGeocodeHandler)
	})

	r.Get("/swagger/*", httpSwagger.WrapHandler)
	fmt.Println("Server is running on port 8080...")
	log.Fatal(http.ListenAndServe(":8080", r))
}